requirements.txt
langchain>=0.3.0
langchain-openai>=0.3.0
langchain-community>=0.3.0
langchain-chroma>=0.2.0
langchain-text-splitters>=0.3.0
chromadb>=0.6.0
openai>=1.60.0
python-dotenv>=1.0.0
streamlit>=1.41.0
tiktoken>=0.8.0
rank-bm25>=0.2.2
